# openapi.yaml
openapi: 3.0.3
info:
  title: Ergonomia API
  description: |
    API REST para la web de afiliados de productos ergonómicos.
    
    Características:
    - Autenticación JWT
    - Gestión de productos
    - Favoritos y comparaciones
    - Admin dashboard
    - HTMX para renderizado parcial
  version: 1.0.0
  contact:
    name: Gentleman Programming
    url: https://github.com/Gentleman-Programming
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.ergonomia.pro
    description: Production server
...
paths:
...
  /api/v1/favorites:
    get:
      tags: [User]
      summary: Listar favoritos
      security:
        - BearerAuth: []
      responses:
        200:
          description: Lista de favoritos

    post:
      tags: [User]
      summary: Añadir a favoritos
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: integer
      responses:
        201:
          description: Favorito añadido
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...